grpc-server/requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.2